+++ src/socket_server/sok_functions/destroy-room.js
@@ -4,10 +4,6 @@ const CONST = require('../consts');
 
 module.exports = (io, ROOMS, sok) => {
    sok.destroyRoom = () => {
-      if (! sok.room) {
-         return;
-      }
-      
       io.to(sok.roomname).emit('chat', sok.username, 'Owner left. Room deleted.');
       
       sok.room.players.forEach(player => {