+++ src/web/js/game/movement.js
@@ -3,9 +3,12 @@
 // returns true if you CANNOT GO through these coordinates (there is a thing at these coords)
 function stop(x, y) {
    return (
-      map[y][x] === BLOCK.NORMAL || map[y][x] === BLOCK.PERMANENT ||
-      bombs.filter(({x: xb, y: yb}) => x === xb && y === yb).length === 1
+      map[y][x] === BLOCK.NORMAL || map[y][x] === BLOCK.PERMANENT || isbomb(x, y)
    );
+}
+
+function isbomb(x, y) {
+   return bombs.filter(({x: xb, y: yb}) => x === xb && y === yb).length === 1;
 }
 
 
@@ -13,9 +16,6 @@ function moveLeft() {
    changeAnimation(myColor, 'left');
    
    if (me.x === MIN_X)
-      return
-
-   if (me.x % BLOCK_SIZE === 0 && me.y % BLOCK_SIZE === 0 && stop(me.x / BLOCK_SIZE - 1, me.y / BLOCK_SIZE))
       return
 
    const mod = me.y % (2 * BLOCK_SIZE)
@@ -38,21 +38,34 @@ function moveLeft() {
       me.x -= moveSpeed * deltaTime
       if (me.x < MIN_X || stop(Math.floor(me.x / BLOCK_SIZE), me.y / BLOCK_SIZE))
          me.x = Math.floor(meOld.x / BLOCK_SIZE) * BLOCK_SIZE
+      if (me.x % BLOCK_SIZE === 0 && isbomb(me.x / BLOCK_SIZE - 1, me.y / BLOCK_SIZE)) {
+         socket.emit('kickbomb', me.x / BLOCK_SIZE - 1, me.y / BLOCK_SIZE, -1, 0);
+      }
    } else {
-      if (mod < BLOCK_SIZE && !stop(me.x / BLOCK_SIZE - 1, Math.floor(me.y / BLOCK_SIZE))) {
-         me.y -= moveSpeed * deltaTime
-         const newMod = me.y % (2 * BLOCK_SIZE)
-         if (newMod > mod) {
-            me.x -= 2 * BLOCK_SIZE - newMod
-            me.y = Math.floor(meOld.y / BLOCK_SIZE) * BLOCK_SIZE
+      if (mod < BLOCK_SIZE) {
+         if (isbomb(me.x / BLOCK_SIZE - 1, Math.floor(me.y / BLOCK_SIZE))) {
+            socket.emit('kickbomb', me.x / BLOCK_SIZE - 1, Math.floor(me.y / BLOCK_SIZE), -1, 0);
+         }
+         if (!stop(me.x / BLOCK_SIZE - 1, Math.floor(me.y / BLOCK_SIZE))) {
+            me.y -= moveSpeed * deltaTime
+            const newMod = me.y % (2 * BLOCK_SIZE)
+            if (newMod > mod) {
+               me.x -= 2 * BLOCK_SIZE - newMod
+               me.y = Math.floor(meOld.y / BLOCK_SIZE) * BLOCK_SIZE
+            }
          }
       }
-      else if (mod > BLOCK_SIZE && !stop(me.x / BLOCK_SIZE - 1, Math.floor(me.y / BLOCK_SIZE) + 1)) {
-         me.y += moveSpeed * deltaTime
-         const newMod = me.y % (2 * BLOCK_SIZE)
-         if (newMod < mod) {
-            me.x -= newMod
-            me.y = Math.floor(me.y / BLOCK_SIZE) * BLOCK_SIZE
+      else if (mod > BLOCK_SIZE) {
+         if (isbomb(me.x / BLOCK_SIZE - 1, Math.floor(me.y / BLOCK_SIZE) + 1)) {
+            socket.emit('kickbomb', me.x / BLOCK_SIZE - 1, Math.floor(me.y / BLOCK_SIZE) + 1, -1, 0);
+         }
+         if (!stop(me.x / BLOCK_SIZE - 1, Math.floor(me.y / BLOCK_SIZE) + 1)) {
+            me.y += moveSpeed * deltaTime
+            const newMod = me.y % (2 * BLOCK_SIZE)
+            if (newMod < mod) {
+               me.x -= newMod
+               me.y = Math.floor(me.y / BLOCK_SIZE) * BLOCK_SIZE
+            }
          }
       }
    }
@@ -66,9 +79,6 @@ function moveDown() {
    changeAnimation(myColor, 'front');
    
    if (me.y === MAX_Y)
-      return
-   
-   if (me.x % BLOCK_SIZE === 0 && me.y % BLOCK_SIZE === 0 && stop(me.x / BLOCK_SIZE, me.y / BLOCK_SIZE + 1))
       return
    
    const mod = me.x % (2 * BLOCK_SIZE)
@@ -90,22 +100,35 @@ function moveDown() {
    if (mod === 0) {
       me.y += moveSpeed * deltaTime
       if (me.y > MAX_Y || stop(me.x / BLOCK_SIZE, Math.floor(me.y / BLOCK_SIZE) + 1))
-         me.y = Math.floor(meOld.y / BLOCK_SIZE + 1) * BLOCK_SIZE
+         me.y = Math.floor(me.y / BLOCK_SIZE) * BLOCK_SIZE
+      if (me.y % BLOCK_SIZE === 0 && isbomb(me.x / BLOCK_SIZE, me.y / BLOCK_SIZE + 1)) {
+         socket.emit('kickbomb', me.x / BLOCK_SIZE, me.y / BLOCK_SIZE + 1, 0, +1);
+      }
    } else {
-      if (mod < BLOCK_SIZE && !stop(Math.floor(me.x / BLOCK_SIZE), me.y / BLOCK_SIZE + 1)) {
-         me.x -= moveSpeed * deltaTime
-         const newMod = me.x % (2 * BLOCK_SIZE)
-         if (newMod > mod) {
-            me.x = Math.floor(meOld.x / BLOCK_SIZE) * BLOCK_SIZE
-            me.y += 2 * BLOCK_SIZE - newMod
+      if (mod < BLOCK_SIZE) {
+         if (isbomb(Math.floor(me.x / BLOCK_SIZE), me.y / BLOCK_SIZE + 1)) {
+            socket.emit('kickbomb', Math.floor(me.x / BLOCK_SIZE), me.y / BLOCK_SIZE + 1, 0, +1);
+         }
+         if (!stop(Math.floor(me.x / BLOCK_SIZE), me.y / BLOCK_SIZE + 1)) {
+            me.x -= moveSpeed * deltaTime
+            const newMod = me.x % (2 * BLOCK_SIZE)
+            if (newMod > mod) {
+               me.x = Math.floor(meOld.x / BLOCK_SIZE) * BLOCK_SIZE
+               me.y += 2 * BLOCK_SIZE - newMod
+            }
          }
       }
-      else if (mod > BLOCK_SIZE && !stop(Math.floor(me.x / BLOCK_SIZE) + 1, me.y / BLOCK_SIZE + 1)) {
-         me.x += moveSpeed * deltaTime
-         const newMod = me.x % (2 * BLOCK_SIZE)
-         if (newMod < mod) {
-            me.x = Math.floor(me.x / BLOCK_SIZE) * BLOCK_SIZE
-            me.y += newMod
+      else if (mod > BLOCK_SIZE) {
+         if (isbomb(Math.floor(me.x / BLOCK_SIZE) + 1, me.y / BLOCK_SIZE + 1)) {
+            socket.emit('kickbomb', Math.floor(me.x / BLOCK_SIZE) + 1, me.y / BLOCK_SIZE + 1, 0, +1);
+         }
+         if (!stop(Math.floor(me.x / BLOCK_SIZE) + 1, me.y / BLOCK_SIZE + 1)) {
+            me.x += moveSpeed * deltaTime
+            const newMod = me.x % (2 * BLOCK_SIZE)
+            if (newMod < mod) {
+               me.x = Math.floor(me.x / BLOCK_SIZE) * BLOCK_SIZE
+               me.y += newMod
+            }
          }
       }
    }
@@ -119,9 +142,6 @@ function moveRight() {
    changeAnimation(myColor, 'right');
    
    if (me.x === MAX_X)
-      return
-   
-   if (me.x % BLOCK_SIZE === 0 && me.y % BLOCK_SIZE === 0 && stop(me.x / BLOCK_SIZE + 1, me.y / BLOCK_SIZE))
       return
    
    const mod = me.y % (2 * BLOCK_SIZE)
@@ -143,22 +163,35 @@ function moveRight() {
    if (mod === 0) {
       me.x += moveSpeed * deltaTime
       if (me.x > MAX_X || stop(Math.floor(me.x / BLOCK_SIZE) + 1, me.y / BLOCK_SIZE))
-         me.x = Math.floor(meOld.x / BLOCK_SIZE + 1) * BLOCK_SIZE
+         me.x = Math.floor(me.x / BLOCK_SIZE) * BLOCK_SIZE
+      if (me.x % BLOCK_SIZE === 0 && isbomb(me.x / BLOCK_SIZE + 1, me.y / BLOCK_SIZE)) {
+         socket.emit('kickbomb', me.x / BLOCK_SIZE + 1, me.y / BLOCK_SIZE, +1, 0);
+      }
    } else {
-      if (mod < BLOCK_SIZE && !stop(me.x / BLOCK_SIZE + 1, Math.floor(me.y / BLOCK_SIZE))) {
-         me.y -= moveSpeed * deltaTime
-         const newMod = me.y % (2 * BLOCK_SIZE)
-         if (newMod > mod) {
-            me.x += 2 * BLOCK_SIZE - newMod
-            me.y = Math.floor(meOld.y / BLOCK_SIZE) * BLOCK_SIZE
+      if (mod < BLOCK_SIZE) {
+         if (isbomb(me.x / BLOCK_SIZE + 1, Math.floor(me.y / BLOCK_SIZE))) {
+            socket.emit('kickbomb', me.x / BLOCK_SIZE + 1, Math.floor(me.y / BLOCK_SIZE), +1, 0);
+         }
+         if (!stop(me.x / BLOCK_SIZE + 1, Math.floor(me.y / BLOCK_SIZE))) {
+            me.y -= moveSpeed * deltaTime
+            const newMod = me.y % (2 * BLOCK_SIZE)
+            if (newMod > mod) {
+               me.x += 2 * BLOCK_SIZE - newMod
+               me.y = Math.floor(meOld.y / BLOCK_SIZE) * BLOCK_SIZE
+            }
          }
       }
-      else if (mod > BLOCK_SIZE && !stop(me.x / BLOCK_SIZE + 1, Math.floor(me.y / BLOCK_SIZE) + 1)) {
-         me.y += moveSpeed * deltaTime
-         const newMod = me.y % (2 * BLOCK_SIZE)
-         if (newMod < mod) {
-            me.x += newMod
-            me.y = Math.floor(me.y / BLOCK_SIZE) * BLOCK_SIZE
+      else if (mod > BLOCK_SIZE) {
+         if (isbomb(me.x / BLOCK_SIZE + 1, Math.floor(me.y / BLOCK_SIZE) + 1)) {
+            socket.emit('kickbomb', me.x / BLOCK_SIZE + 1, Math.floor(me.y / BLOCK_SIZE) + 1, +1, 0);
+         }
+         if (!stop(me.x / BLOCK_SIZE + 1, Math.floor(me.y / BLOCK_SIZE) + 1)) {
+            me.y += moveSpeed * deltaTime
+            const newMod = me.y % (2 * BLOCK_SIZE)
+            if (newMod < mod) {
+               me.x += newMod
+               me.y = Math.floor(me.y / BLOCK_SIZE) * BLOCK_SIZE
+            }
          }
       }
    }
@@ -172,9 +205,6 @@ function moveUp() {
    changeAnimation(myColor, 'back');
 
    if (me.y === MIN_Y)
-      return
-   
-   if (me.x % BLOCK_SIZE === 0 && me.y % BLOCK_SIZE === 0 && stop(me.x / BLOCK_SIZE, me.y / BLOCK_SIZE - 1))
       return
    
    const mod = me.x % (2 * BLOCK_SIZE)
@@ -197,21 +227,34 @@ function moveUp() {
       me.y -= moveSpeed * deltaTime
       if (me.y < MIN_Y || stop(me.x / BLOCK_SIZE, Math.floor(me.y / BLOCK_SIZE)))
          me.y = Math.floor(meOld.y / BLOCK_SIZE) * BLOCK_SIZE
+      if (me.y % BLOCK_SIZE === 0 && isbomb(me.x / BLOCK_SIZE, me.y / BLOCK_SIZE - 1)) {
+         socket.emit('kickbomb', me.x / BLOCK_SIZE, me.y / BLOCK_SIZE - 1, 0, -1);
+      }
    } else {
-      if (mod < BLOCK_SIZE && !stop(Math.floor(me.x / BLOCK_SIZE), me.y / BLOCK_SIZE - 1)) {
-         me.x -= moveSpeed * deltaTime
-         const newMod = me.x % (2 * BLOCK_SIZE)
-         if (newMod > mod) {
-            me.x = Math.floor(meOld.x / BLOCK_SIZE) * BLOCK_SIZE
-            me.y -= 2 * BLOCK_SIZE - newMod
+      if (mod < BLOCK_SIZE) {
+         if (isbomb(Math.floor(me.x / BLOCK_SIZE), me.y / BLOCK_SIZE - 1)) {
+            socket.emit('kickbomb', Math.floor(me.x / BLOCK_SIZE), me.y / BLOCK_SIZE - 1, 0, -1);
+         }
+         if (!stop(Math.floor(me.x / BLOCK_SIZE), me.y / BLOCK_SIZE - 1)) {
+            me.x -= moveSpeed * deltaTime
+            const newMod = me.x % (2 * BLOCK_SIZE)
+            if (newMod > mod) {
+               me.x = Math.floor(meOld.x / BLOCK_SIZE) * BLOCK_SIZE
+               me.y -= 2 * BLOCK_SIZE - newMod
+            }
          }
       }
-      else if (mod > BLOCK_SIZE && !stop(Math.floor(me.x / BLOCK_SIZE) + 1, me.y / BLOCK_SIZE - 1)) {
-         me.x += moveSpeed * deltaTime
-         const newMod = me.x % (2 * BLOCK_SIZE)
-         if (newMod < mod) {
-            me.x = Math.floor(me.x / BLOCK_SIZE) * BLOCK_SIZE
-            me.y -= newMod
+      else if (mod > BLOCK_SIZE) {
+         if (isbomb(Math.floor(me.x / BLOCK_SIZE) + 1, me.y / BLOCK_SIZE - 1)) {
+            socket.emit('kickbomb', Math.floor(me.x / BLOCK_SIZE) + 1, me.y / BLOCK_SIZE - 1, 0, -1);
+         }
+         if (!stop(Math.floor(me.x / BLOCK_SIZE) + 1, me.y / BLOCK_SIZE - 1)) {
+            me.x += moveSpeed * deltaTime
+            const newMod = me.x % (2 * BLOCK_SIZE)
+            if (newMod < mod) {
+               me.x = Math.floor(me.x / BLOCK_SIZE) * BLOCK_SIZE
+               me.y -= newMod
+            }
          }
       }
    }